modify some pages
